// src/interfaces/movie.interface.ts
export interface Movie {
  id: number;
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface MovieData {
  results: Movie[];
  page: number;
  total_pages: number;
  total_results: number;
  dates: { maximun: string; minimum: string };
}